requirements.txt
fastapi
uvicorn
playwright
pandas
requests
openai
httpx
python-dotenv
beautifulsoup4
pydantic
